agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "FastAPI agent server for Octopets chat experience"
readme = "README.md"
authors = [
    { name = "Octopets Team" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.15.0",
    "agent-framework[foundry]",
    "opentelemetry-api>=1.33.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.33.0",
    "opentelemetry-instrumentation-fastapi>=0.54b0",
    "opentelemetry-sdk>=1.33.0",
    "grpcio>=1.50.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.uv.sources]
"agent-framework" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/main" }
"agent-framework-azure" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/azure" }
"agent-framework-foundry" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/foundry" }
"agent-framework-workflow" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/workflow" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
